Implementacion formularios
lote, remision, inventario, referencia producto

+++ OpenMind/src/Controlador/OpmRemisionJpaController.java
@@ -5,16 +5,19 @@
  */
 package Controlador;
 
+import Controlador.exceptions.IllegalOrphanException;
 import Controlador.exceptions.NonexistentEntityException;
-import Modelo.OpmRemision;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import Modelo.OpmDetalleRemision;
+import Modelo.OpmRemision;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -32,11 +35,29 @@ public class OpmRemisionJpaController implements Serializable {
     }
 
     public void create(OpmRemision opmRemision) {
+        if (opmRemision.getOpmDetalleRemisionList() == null) {
+            opmRemision.setOpmDetalleRemisionList(new ArrayList<OpmDetalleRemision>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<OpmDetalleRemision> attachedOpmDetalleRemisionList = new ArrayList<OpmDetalleRemision>();
+            for (OpmDetalleRemision opmDetalleRemisionListOpmDetalleRemisionToAttach : opmRemision.getOpmDetalleRemisionList()) {
+                opmDetalleRemisionListOpmDetalleRemisionToAttach = em.getReference(opmDetalleRemisionListOpmDetalleRemisionToAttach.getClass(), opmDetalleRemisionListOpmDetalleRemisionToAttach.getNmCodigo());
+                attachedOpmDetalleRemisionList.add(opmDetalleRemisionListOpmDetalleRemisionToAttach);
+            }
+            opmRemision.setOpmDetalleRemisionList(attachedOpmDetalleRemisionList);
             em.persist(opmRemision);
+            for (OpmDetalleRemision opmDetalleRemisionListOpmDetalleRemision : opmRemision.getOpmDetalleRemisionList()) {
+                OpmRemision oldNmRemisionOfOpmDetalleRemisionListOpmDetalleRemision = opmDetalleRemisionListOpmDetalleRemision.getNmRemision();
+                opmDetalleRemisionListOpmDetalleRemision.setNmRemision(opmRemision);
+                opmDetalleRemisionListOpmDetalleRemision = em.merge(opmDetalleRemisionListOpmDetalleRemision);
+                if (oldNmRemisionOfOpmDetalleRemisionListOpmDetalleRemision != null) {
+                    oldNmRemisionOfOpmDetalleRemisionListOpmDetalleRemision.getOpmDetalleRemisionList().remove(opmDetalleRemisionListOpmDetalleRemision);
+                    oldNmRemisionOfOpmDetalleRemisionListOpmDetalleRemision = em.merge(oldNmRemisionOfOpmDetalleRemisionListOpmDetalleRemision);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +66,45 @@ public class OpmRemisionJpaController implements Serializable {
         }
     }
 
-    public void edit(OpmRemision opmRemision) throws NonexistentEntityException, Exception {
+    public void edit(OpmRemision opmRemision) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            OpmRemision persistentOpmRemision = em.find(OpmRemision.class, opmRemision.getNmCodigo());
+            List<OpmDetalleRemision> opmDetalleRemisionListOld = persistentOpmRemision.getOpmDetalleRemisionList();
+            List<OpmDetalleRemision> opmDetalleRemisionListNew = opmRemision.getOpmDetalleRemisionList();
+            List<String> illegalOrphanMessages = null;
+            for (OpmDetalleRemision opmDetalleRemisionListOldOpmDetalleRemision : opmDetalleRemisionListOld) {
+                if (!opmDetalleRemisionListNew.contains(opmDetalleRemisionListOldOpmDetalleRemision)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain OpmDetalleRemision " + opmDetalleRemisionListOldOpmDetalleRemision + " since its nmRemision field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<OpmDetalleRemision> attachedOpmDetalleRemisionListNew = new ArrayList<OpmDetalleRemision>();
+            for (OpmDetalleRemision opmDetalleRemisionListNewOpmDetalleRemisionToAttach : opmDetalleRemisionListNew) {
+                opmDetalleRemisionListNewOpmDetalleRemisionToAttach = em.getReference(opmDetalleRemisionListNewOpmDetalleRemisionToAttach.getClass(), opmDetalleRemisionListNewOpmDetalleRemisionToAttach.getNmCodigo());
+                attachedOpmDetalleRemisionListNew.add(opmDetalleRemisionListNewOpmDetalleRemisionToAttach);
+            }
+            opmDetalleRemisionListNew = attachedOpmDetalleRemisionListNew;
+            opmRemision.setOpmDetalleRemisionList(opmDetalleRemisionListNew);
             opmRemision = em.merge(opmRemision);
+            for (OpmDetalleRemision opmDetalleRemisionListNewOpmDetalleRemision : opmDetalleRemisionListNew) {
+                if (!opmDetalleRemisionListOld.contains(opmDetalleRemisionListNewOpmDetalleRemision)) {
+                    OpmRemision oldNmRemisionOfOpmDetalleRemisionListNewOpmDetalleRemision = opmDetalleRemisionListNewOpmDetalleRemision.getNmRemision();
+                    opmDetalleRemisionListNewOpmDetalleRemision.setNmRemision(opmRemision);
+                    opmDetalleRemisionListNewOpmDetalleRemision = em.merge(opmDetalleRemisionListNewOpmDetalleRemision);
+                    if (oldNmRemisionOfOpmDetalleRemisionListNewOpmDetalleRemision != null && !oldNmRemisionOfOpmDetalleRemisionListNewOpmDetalleRemision.equals(opmRemision)) {
+                        oldNmRemisionOfOpmDetalleRemisionListNewOpmDetalleRemision.getOpmDetalleRemisionList().remove(opmDetalleRemisionListNewOpmDetalleRemision);
+                        oldNmRemisionOfOpmDetalleRemisionListNewOpmDetalleRemision = em.merge(oldNmRemisionOfOpmDetalleRemisionListNewOpmDetalleRemision);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +122,7 @@ public class OpmRemisionJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +133,17 @@ public class OpmRemisionJpaController implements Serializable {
                 opmRemision.getNmCodigo();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The opmRemision with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<OpmDetalleRemision> opmDetalleRemisionListOrphanCheck = opmRemision.getOpmDetalleRemisionList();
+            for (OpmDetalleRemision opmDetalleRemisionListOrphanCheckOpmDetalleRemision : opmDetalleRemisionListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This OpmRemision (" + opmRemision + ") cannot be destroyed since the OpmDetalleRemision " + opmDetalleRemisionListOrphanCheckOpmDetalleRemision + " in its opmDetalleRemisionList field has a non-nullable nmRemision field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(opmRemision);
             em.getTransaction().commit();
